Split PCXEditorCanvas into the editor and the visualizer.
